Introduce CallbackType in Scxml entries for enforcing ROS events in expressions
